requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
Flask==3.1.0
fonttools==4.57.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.6.3
numpy==1.24.1
opencv-python==4.7.0.68
openpyxl==3.0.10
packaging==25.0
pandas==1.5.2
Pillow==9.4.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tqdm==4.64.1
tzdata==2025.2
websockets==11.0.3
Werkzeug==3.1.3
